refactor(cua-driver): test-suite taxonomy, protocol split+dedup, transport test (Phases 3-6)
Phase 3 — rename tests into a 4-family taxonomy (git mv, blame preserved):
  protocol_ / transport_ / harness_ / modality_ / guard_.
  harness_web_windows→harness_web, harness_lo_vcl→harness_libreoffice,
  harness_bg_modality→modality_background, harness_desktop_scope→modality_desktop_scope,
  focus_check→modality_focus, e2e_windows_bg_input→modality_input_e2e,
  ux_guard→guard_ux, element_token→protocol_element_token.

Phase 5 — split mcp_protocol_test (3,412 lines / 65 tests) into 5 protocol_*_test
  files (handshake/tools_call/schema/media/session) on a new testkit RawDriver
  (raw send/recv, no auto-init). Deduped 24 macOS↔Windows mirror pairs into single
  cfg!-branching tests; 13 mac-only + 4 win-only kept. All 40 macOS protocol tests
  pass post-split.

Phase 4 — add transport_config_persistence_test: set_config persists to disk
  across stateless CLI invocations (#2034) vs visible within an MCP session — the
  one behavior only observable across both transports.

Phase 6 — rewrite TEST_SUITE.md / TEST_HARNESS_STRUCTURE.md to the new names +
  a transport × modality × platform coverage matrix.

Deferred (noted): merging modality_input_e2e ↔ modality_background, and
extracting the focus-steal sentinel into a shared testkit assert.
